docshell/test/mochitest/test_load_during_reload.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Test loading a new page after calling reload()</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" href="/tests/SimpleTest/test.css"/>
  <script>

    function promiseForLoad() {
      return new Promise(resolve => {
        addEventListener("message", resolve, { once: true });
      });
    }

    add_task(async function runTest() {
      let win = window.open("file_load_during_reload.html");
      await promiseForLoad();

      win.location.reload();
      win.location.href = "file_load_during_reload.html?nextpage";
      await promiseForLoad();

      ok(win.location.href.includes("nextpage"), "Should have loaded the next page.");
      win.close();
    });

  </script>
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none"></div>
<pre id="test"></pre>
</body>
</html>
